requirements.txt
asgiref==3.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.5
commonmark==0.9.1
cv==1.0.0
Cython==0.29.24
defusedxml==0.7.1
Django==3.2.13
django-cors-headers==3.13.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.0
ffmpeg==1.4
httpie==3.2.1
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.3
multidict==6.0.2
npm==0.1.1
opencv-contrib-python==4.5.3.56
optional-django==0.1.0
Pillow==9.1.1
pydub-ffmpeg-wav==0.23.0
Pygments==2.12.0
PyJWT==2.4.0
pyserial==3.5
PySocks==1.7.1
pytz==2021.3
pyworld==0.2.12
requests==2.28.0
requests-toolbelt==0.9.1
rich==12.4.4
typing_extensions==4.2.0
urllib3==1.26.9
wincertstore==0.2
xlrd==1.2.0
you-get==0.4.1527
zipp==3.8.0
